feat(module): :sparkles: New use case "view all" and fix bugs

+++ src/api/v1/inventory/infrastructure/http/controllers/fastapi_inventory_controller.py
@@ -1,15 +1,18 @@
+from typing import List
 from src.api.v1.shared.domain.value_objects import Uuid
 from fastapi import HTTPException
 from src.api.v1.inventory.domain.errors import (
     InventoryItemTypeError,
     InventoryItemError,
 )
-from src.api.v1.shared.domain.value_objects.uuid import Uuid
-from src.api.v1.inventory.infrastructure.persistence.models import (
+from src.api.v1.inventory.infrastructure.persistence.models.sqlmodel_inventory_model import (  # noqa: E501
     SqlModelInventoryModel,
 )
 from src.api.v1.inventory.infrastructure.persistence.repositories import (
     SQLModelInventoryRepository,
+)
+from src.api.v1.user.infrastructure.persistence.repositories.sqlmodel_user_repository import (  # noqa: E501
+    SqlModelUserRepository,
 )
 from src.api.v1.inventory.infrastructure.http.dtos import (
     PydanticCreateItemRequestDto,
@@ -20,10 +23,19 @@ from src.api.v1.inventory.infrastructure.http.dtos import (
     PydanticViewItemResponseDto,
 )
 from src.api.v1.inventory.application.create import CreateItem
-from src.api.v1.inventory.application.delete import DeleteItem, DeleteItemDTO
+from src.api.v1.inventory.application.delete import DeleteItem
+from src.api.v1.inventory.application.delete.delete_item_dto import DeleteItemDTO
 from src.api.v1.inventory.application.update import UpdateItem
-from src.api.v1.inventory.application.view import ViewItem, ViewItemDTO
+from src.api.v1.inventory.application.view import ViewItem
+from src.api.v1.inventory.application.view.view_item_dto import ViewItemDTO
+from src.api.v1.inventory.application.view_all import ViewAllInventoryItems
+from src.api.v1.inventory.application.view_all.view_all_item_dto import (
+    ViewAllInventoryItemsDTO,
+)
 from src.api.v1.shared.infrastructure.persistence.sqlmodel_connection import get_session
+from src.api.v1.user.infrastructure.http.services.in_memory_session_service import (
+    InMemorySessionService,
+)
 from .exeption_handler import handle_exceptions
 
 
@@ -35,9 +47,13 @@ class FastApiInventoryController:
     ) -> PydanticCreateItemResponseDto:
         with get_session() as session:
             repo = SQLModelInventoryRepository(session=session)
-            use_case = CreateItem(repo)
+            user_repo = SqlModelUserRepository(db_connection=session)
+            InMemorySessionService.validate_permission(
+                Uuid(user_id), Uuid(item_data.user_id)
+            )
+
+            use_case = CreateItem(repo, user_repo)
             dto = item_data.to_application()
-            logging.debug("DEBUG - DTO passed to use case: %s", dto)
             item = use_case.execute(dto)
             return PydanticCreateItemResponseDto(
                 item=SqlModelInventoryModel.from_entity(item)
@@ -50,15 +66,25 @@ class FastApiInventoryController:
     ) -> PydanticUpdateItemResponseDto:
         with get_session() as session:
             repo = SQLModelInventoryRepository(session=session)
-            use_case = UpdateItem(repo)
-            dto = item_data.to_application()
-            use_case.execute(dto)
-            updated_item = repo.find_by_id(Uuid(dto.inventory_id))
-            if not updated_item:
+            inventory_item = repo.find_by_id(Uuid(item_data.inventory_id))
+            if not inventory_item:
                 raise HTTPException(
                     status_code=404,
                     detail=InventoryItemError(InventoryItemTypeError.ITEM_NOT_FOUND),
                 )
+            InMemorySessionService.validate_permission(
+                Uuid(user_id), inventory_item.user_id
+            )
+
+            use_case = UpdateItem(repo)
+            dto = item_data.to_application()
+            use_case.execute(dto)
+            updated_item = repo.find_by_id(Uuid(dto.inventory_id))
+            if updated_item is None:
+                raise HTTPException(
+                    status_code=404, detail="Updated inventory item not found"
+                )
+
             return PydanticUpdateItemResponseDto(
                 item=SqlModelInventoryModel.from_entity(updated_item)
             )
@@ -68,6 +94,16 @@ class FastApiInventoryController:
     async def delete(inventory_id: str, user_id: str) -> PydanticDeleteItemResponseDto:
         with get_session() as session:
             repo = SQLModelInventoryRepository(session=session)
+            inventory_item = repo.find_by_id(Uuid(inventory_id))
+            if not inventory_item:
+                raise HTTPException(
+                    status_code=404,
+                    detail=InventoryItemError(InventoryItemTypeError.ITEM_NOT_FOUND),
+                )
+            InMemorySessionService.validate_permission(
+                Uuid(user_id), inventory_item.user_id
+            )
+
             use_case = DeleteItem(repo)
             dto = DeleteItemDTO(inventory_id=inventory_id)
             deleted_item = use_case.execute(dto)
@@ -80,14 +116,38 @@ class FastApiInventoryController:
     async def view(inventory_id: str, user_id: str) -> PydanticViewItemResponseDto:
         with get_session() as session:
             repo = SQLModelInventoryRepository(session=session)
-            use_case = ViewItem(repo)
-            dto = ViewItemDTO(inventory_id=inventory_id)
-            item = use_case.execute(dto)
-            if not item:
+            inventory_item = repo.find_by_id(Uuid(inventory_id))
+            if not inventory_item:
                 raise HTTPException(
                     status_code=404,
                     detail=InventoryItemError(InventoryItemTypeError.ITEM_NOT_FOUND),
                 )
+            InMemorySessionService.validate_permission(
+                Uuid(user_id), inventory_item.user_id
+            )
+
+            use_case = ViewItem(repo)
+            dto = ViewItemDTO(inventory_id=inventory_id)
+            item = use_case.execute(dto)
             return PydanticViewItemResponseDto(
                 item=SqlModelInventoryModel.from_entity(item)
             )
+
+    @staticmethod
+    async def view_all(user_id: str) -> List[PydanticViewItemResponseDto]:
+        with get_session() as session:
+            repo = SQLModelInventoryRepository(session=session)
+            use_case = ViewAllInventoryItems(repo)
+
+            dto = ViewAllInventoryItemsDTO(user_id=Uuid(user_id))
+            inventory_items = use_case.execute(dto)
+
+            if not inventory_items:
+                raise HTTPException(status_code=404, detail="No inventory items found")
+
+            return [
+                PydanticViewItemResponseDto(
+                    item=SqlModelInventoryModel.from_entity(item)
+                )
+                for item in inventory_items
+            ]
